+++ R/explore.R
@@ -446,8 +446,7 @@ explore <- function(
                                     bio = bio)
 
     if (any(sapply(valid.detections, function(x) any(!is.na(x$Sensor.Value))))) {
-      appendTo(c("Screen", "Report"), "M: Printing sensor values for tags with sensor data.")
-      sensor.plots <- printSensorData(detections = valid.detections)
+      sensor.plots <- printSensorData(detections = valid.detections, rsp.info = rsp.info)
     } else {
       sensor.plots <- NULL
     }